requirements.txt
streamlit
pymupdf
pandas
openpyxl
google-genai